refactor(backend): refactor user authentification

+++ packages/backend/src/_migrations/1741040395992-update-open-data-places-migration.ts
@@ -1,16 +1,23 @@
 import { MigrationInterface, QueryRunner } from "typeorm";
+import { domifaConfig } from "../config";
 
 export class AutoMigration1741040395992 implements MigrationInterface {
   name = "AutoMigration1741040395992";
 
   public async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query(
-      `ALTER TABLE open_data_places DROP COLUMN "uniqueId"`
-    );
+    if (
+      domifaConfig().envId === "prod" ||
+      domifaConfig().envId === "preprod" ||
+      domifaConfig().envId === "local"
+    ) {
+      await queryRunner.query(
+        `ALTER TABLE open_data_places DROP COLUMN "uniqueId"`
+      );
 
-    await queryRunner.query(
-      `DELETE FROM open_data_places where source != 'dgcs' `
-    );
+      await queryRunner.query(
+        `DELETE FROM open_data_places where source != 'dgcs' `
+      );
+    }
   }
 
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
